Move files to own facets folder
